SourceMap shall also work when merging multiple input files with a source map for each.
SourceMap now accepts a function as options.orig.
In case options.orig is a string, SourceMapConsumer is set up like before.
In case options.orig is a function, for each source file this function is invoked.
The function needs to be (string) => string|Object. For a source file name it shall return the contents of the original map for this source file.
The result of the function call is cached in a Map [sourceFileName] => SourceMapConsumer, such that for each source file the SourceMapConsumer is generated only once.